Creación de Custom Message Boxes y otros cambios

+++ SupportHub/SupportHub/frmEntrega.cs
@@ -1,5 +1,6 @@
 ﻿using Comun.Cache;
 using Dominio;
+using Presentacion.CustomMessageBoxes;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -38,9 +39,8 @@ namespace Presentacion
 
         private void frmEntrega_Load(object sender, EventArgs e)
         {
-            prepararDgvEntregas();
-
             formCargado = true;
+            prepararDgvEntregas();
 
             if (!CacheInicioUsuario.permisosUser.Contains("Agregar"))
             {
@@ -136,9 +136,9 @@ namespace Presentacion
 
         private void btnEliminarEntrega_Click(object sender, EventArgs e)
         {
-            DialogResult resultado = MessageBox.Show("¿Seguro que desea eliminar entregas?", "Eliminar entregas", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            AdvertenciaMessageBox advertencia = new("Confirmar eliminación","¿Seguro que desea eliminar entregas?");
 
-            if (resultado == DialogResult.Yes)
+            if (advertencia.obtenerConfirmacion() == DialogResult.Yes)
             {
                 if (dgvEntregas.SelectedRows.Count > 0)
                 {
@@ -152,16 +152,16 @@ namespace Presentacion
 
                     if (totalRegistrosEliminados > 1)
                     {
-                        MessageBox.Show("Se eliminaron " + totalRegistrosEliminados.ToString() + " entregas.","Eliminación exitosa", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        ExitoMessageBox exito = new("Eliminación exitosa", "Se eliminaron " + totalRegistrosEliminados.ToString() + " entregas.");
                     }
                     else
                     {
-                        MessageBox.Show("Se eliminó " + totalRegistrosEliminados.ToString() + " entrega.", "Eliminación exitosa", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        ExitoMessageBox exito = new("Eliminación exitosa", "Se eliminó " + totalRegistrosEliminados.ToString() + " entrega.");
                     }
                 }
                 else
                 {
-                    MessageBox.Show("Debe seleccionar una o varias entregas para eliminarlas.", "Error en selección", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    ErrorMessageBox error = new("Error en selección", "Debe seleccionar una o varias entregas para eliminarlas.");
                 }
                 actualizarTablaEntregas();
             }
@@ -171,7 +171,7 @@ namespace Presentacion
         {
             if (CacheInicioUsuario.idEmpleado == 0)
             {
-                MessageBox.Show("Su usuario debe tener un empleado asociado para poder agregar una entrega.", "Usuario Inválido", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                ErrorMessageBox error = new("Usuario Inválido", "Su usuario debe tener un empleado asociado para poder agregar una entrega.");
             }
             else
             {
@@ -213,7 +213,11 @@ namespace Presentacion
             }
             else if (btnGenerarDevolucion.Text == "ELIMINAR DEVOLUCIÓN")
             {
-                devolucionEntrega.eliminarFechaDevolucion();
+                AdvertenciaMessageBox advertencia = new("Eliminar devolución", $"¿Está seguro que desea eliminar la fecha de devolución de la entrega {codEntrega}?");
+                if (advertencia.obtenerConfirmacion() == DialogResult.Yes)
+                {
+                    devolucionEntrega.eliminarFechaDevolucion();
+                }
             }
         }
 
@@ -261,11 +265,12 @@ namespace Presentacion
                     }
                 }
             }
-            else if (formCargado && dgvEntregas.SelectedRows.Count > 1)
+            else if (formCargado && dgvEntregas.SelectedRows.Count != 1)
             {
                 btnGenerarDevolucion.Text = "DEVOLUCIÓN";
                 btnGenerarDevolucion.IconChar = FontAwesome.Sharp.IconChar.RotateBack;
                 btnGenerarDevolucion.Enabled = false;
+
                 btnModificarEntrega.Enabled = false;
             }
         }
